requirements/dev.txt
accessify==0.3.1
azure-cognitiveservices-speech==1.15.0
boto3==1.16.18
botocore==1.19.18
python-dotenv==0.15.0
torch==1.7.0
transformers==4.2.2
spacy==3.0.1
mypy==0.812
pylint==2.7.4
black==20.8b1
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.0.0/en_core_web_md-3.0.0-py3-none-any.whl
click==7.1.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844